% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anomalies}
\alias{anomalies}
\alias{anomaly1}
\alias{anomaly2}
\alias{anomaly3}
\alias{anomaly4}
\alias{anomaly5}
\title{Data sets with anomalies}
\format{
A datasets with anomalies
}
\source{
Created by Di Cook.
}
\usage{
anomaly1

anomaly2

anomaly3

anomaly4

anomaly5
}
\description{
Simulated data with anomalies
}
\details{
\describe{
\item{x1, x2, x3, x4}{numeric variables}
}
}
\examples{
require(GGally)
data(anomaly1)
data(anomaly2)
ggscatmat(anomaly1)
ggscatmat(anomaly2)
}
\keyword{datasets}
