% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anomaly1}
\alias{anomaly1}
\alias{anomaly2,}
\alias{anomaly3,}
\alias{anomaly4,}
\alias{anomaly5}
\title{Data sets with anomalies}
\format{
A datasets with anomalies
}
\description{
Simulated data with anomalies
}
\details{
\describe{
\item{x1, x2, x3, x4}{numeric variables}
}
}
\examples{
require(GGally)
ggscatmat(anomaly1)
ggscatmat(anomaly2)
}
\keyword{datasets}
