% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assoc1}
\alias{assoc1}
\alias{assoc2,}
\alias{assoc3}
\title{Data sets with different types of association}
\format{
A datasets with various association
}
\description{
Simulated data with various associations
}
\details{
\describe{
\item{x1, x2, x3, x4}{numeric variables}
}
}
\examples{
require(GGally)
ggscatmat(assoc1)
}
\keyword{datasets}
