% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{box}
\alias{box}
\title{3D plane in 5D}
\format{
A
}
\description{
This data is simulated to use for testing.
It has three dimensions of variability and
two of noise. It is created from a 3 factor
model. All variables are linearly associated.
}
\details{
\itemize{
x1, x2, x3, x4, x5 are five variables
}
}
\examples{
box_pca <- prcomp(box)
ggscree(box_pca)
}
\seealso{
plane
}
\keyword{datasets}
