% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c1}
\alias{c1}
\alias{c2}
\alias{c3}
\alias{c4}
\alias{c5}
\alias{c6}
\alias{c7}
\title{Challenge data sets}
\format{
A datasets with differing number of rows and columns
}
\description{
Simulated data with different structure
}
\details{
\describe{
\item{x1, x2, ... }{numeric variables}
}
}
\examples{
require(ggplot2)
ggplot(c1, aes(x=x1, y=x2)) +
  geom_point() + theme(aspect.ratio=1)
}
\keyword{datasets}
