% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{calc_norm}
\alias{calc_norm}
\title{Calculate the norm of a vector}
\usage{
calc_norm(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Returns the square root of the sum of squares
of a vector
}
\examples{
x <- rnorm(5)
calc_norm(x)
}
