% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusterchallenges}
\alias{clusterchallenges}
\alias{c1}
\alias{c2}
\alias{c3}
\alias{c4}
\alias{c5}
\alias{c6}
\alias{c7}
\title{Cluster challenge data sets}
\format{
A datasets with differing number of rows and columns
}
\source{
Created by Di Cook.
}
\usage{
c1

c2

c3

c4

c5

c6

c7
}
\description{
Simulated data with different structures
}
\details{
\describe{
\item{x1, x2, ... }{numeric variables}
}
}
\examples{
require(ggplot2)
data(c1)
ggplot(c1, aes(x=x1, y=x2)) +
  geom_point() + theme(aspect.ratio=1)
}
\keyword{datasets}
