% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusters}
\alias{clusters}
\title{Three clusters in 5D}
\format{
A
}
\description{
This data is simulated to use for testing.
It has three elliptical clusters in mostly
variables 2 and 4. They are not equidistant.
}
\details{
\itemize{
x1, x2, x3, x4, x5 are five variables
}
}
\examples{
clusters_pca <- prcomp(clusters)
ggscree(clusters_pca)
}
\seealso{
simple_clusters
}
\keyword{datasets}
