% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{convert_proj_tibble}
\alias{convert_proj_tibble}
\title{This function turns a projection sequence into a tibble}
\usage{
convert_proj_tibble(t1)
}
\arguments{
\item{t1}{tour projection sequence}
}
\description{
Take an array of a projection sequence, and turn into
a tibble with numbered projections
}
\examples{

library(tourr)
t1 <- interpolate(save_history(flea[, 1:6], grand_tour(4), max = 2))
tbl1 <- convert_proj_tibble(t1)
}
