% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{f_mv_dist}
\alias{f_mv_dist}
\title{Compute Mahalanobis distances between all pairs of observations}
\usage{
f_mv_dist(x)
}
\arguments{
\item{x}{multivariate data set}
}
\value{
vector of length n
}
\description{
For a data matrix, compute the sample variance-covariance,
which is used to compute the Mahalanobis distance.
}
\details{
This is useful for checking distance arise from a
multivariate normal sample.
}
\examples{
cat("make an example\n")
}
