% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.R
\name{f_som}
\alias{f_som}
\title{Process the output from SOM to display the map and data}
\usage{
f_som(x, x_som)
}
\arguments{
\item{x}{multivariate data set}

\item{x_som}{object returned by som}
}
\description{
This function generates a grid of points to match the
nodes from the self-organising map (SOM), and jitters points
from the data so they can be seen relative to the grid.
This allows the clustering of points by SOM to be inspected.
}
\examples{
cat("make an example\n")
}
