% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.R
\name{f_som_net}
\alias{f_som_net}
\title{Generate the SOM net}
\usage{
f_som_net(x_som)
}
\arguments{
\item{x_som}{object returned by som}
}
\description{
Generates the edge set from an SOM model.
Combine this with the data matrix to display
the model in p-dimensional space.
}
\examples{
cat("make an example\n")
}
