% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{f_var_ellipse}
\alias{f_var_ellipse}
\title{Ellipse matching data center and variance}
\usage{
f_var_ellipse(x, n = 100)
}
\arguments{
\item{x}{multivariate data set.}

\item{n}{number of points to generate}
}
\value{
matrix of size n x p
}
\description{
This function generates points on the surface of an
ellipse with the same center and variance-covariance
of the provided data.
}
\details{
This is useful for checking the equal variance-covariance
assumption from linear discriminant analysis.
}
\examples{
cat("make an example\n")
}
