% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{gen_chull}
\alias{gen_chull}
\title{Generate a convex hull around clusters to be added to data}
\usage{
gen_chull(data, cl)
}
\arguments{
\item{data}{data set}

\item{cl}{numeric vector of cluster labels}
}
\value{
list with convex hull containing points and edges
}
\description{
Supplements a data set with information needed to draw a
convex hull.
}
\examples{
data(clusters)
cl_dist <- dist(clusters[,1:5])
cl_hw <- hclust(cl_dist, method="ward.D2")
cl_w <- cutree(cl_hw, 3)
cl_chull <- gen_chull(clusters[,1:5], cl_w)
if (interactive()) {
  require(tourr)
  animate_xy(cl_chull$data[,1:5],
    col=cl_chull$data[,6],
    edges=cl_chull$edges,
    edges.col=cl_chull$edge_clr,
    axes="bottomleft")
}
}
\keyword{cluster}
