% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{ggscree}
\alias{ggscree}
\title{This function produces a simple scree plot}
\usage{
ggscree(pc)
}
\arguments{
\item{pc}{PCA object}
}
\value{
scree a ggplot object
}
\description{
Takes a PCA object returned by \code{prcomp()}, extracts
the standard deviations of the principal components (PC), and
plots these against the PC number.
}
\examples{
data(aflw)
pca <- prcomp(aflw[,7:10])
scree <- ggscree(pca)
scree
}
