% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mulgar-package.R
\docType{package}
\name{mulgar-package}
\alias{mulgar}
\alias{mulgar-package}
\title{mulgar: Functions for multivaariaate dat visualisation}
\description{
This package is a companion to the book Interactive and dynamic graphics for multivariate data in R by Cook, Laa, Lee and Wang.
}
\author{
\strong{Maintainer}: Dianne Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})

Other contributors:
\itemize{
  \item Stuart Lee \email{stuart.lee1@monash.edu} (\href{https://orcid.org/0000-0003-1179-8436}{ORCID}) [contributor]
  \item Ursula Laa \email{ursula.laa@boku.ac.at} (\href{https://orcid.org/0000-0002-0249-6439}{ORCID}) [contributor]
  \item Earo Wang \email{earo.wang@gmail.com} (\href{https://orcid.org/0000-0001-6448-5260}{ORCID}) [contributor]
}

}
\keyword{internal}
