% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{norm_vec}
\alias{norm_vec}
\title{Normalise a vector to have length 1}
\usage{
norm_vec(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Returns the normalised vector, where the sum of
squares is equal to 1
}
