% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PISA test scores}
\alias{PISA test scores}
\alias{pisa}
\title{PISA scores}
\format{
A
}
\description{
This is data from the 2018 testing, available from
https://webfs.oecd.org/pisa2018/SPSS_STU_QQQ.zip.
A subset of the data has been generated using only
Australian scores, and the simulated scores for
math, reading and science.
}
\details{
\itemize{
CNT Country (AUS, IDN)
PV1MATH-PV10SCIE simulated scores
}
}
\examples{

require(dplyr)
data(pisa)
pisa \%>\% count(CNT)
}
\keyword{datasets}
