% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pisa}
\alias{pisa}
\title{PISA scores}
\format{
A data set with 26371 rows and 31 columns
}
\description{
This is data from the 2018 testing, available from
https://webfs.oecd.org/pisa2018/SPSS_STU_QQQ.zip.
A subset of the data containing only
Australia and Indonesia, and the simulated
scores for math, reading and science.
}
\details{
\itemize{
\item CNT Country (Australia, Indonesia)
\item PV1MATH-PV10SCIE simulated scores for math, reading and science
}
}
\examples{

require(dplyr)
data(pisa)
pisa \%>\% count(CNT)
}
\keyword{datasets}
