% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_clusters}
\alias{simple_clusters}
\title{Two clusters in 2D}
\format{
A
}
\description{
This data is simulated to use for testing.
It has two spherical clusters, and two variables.
}
\details{
\itemize{
x1, x2 are variables
}
}
\examples{
require(ggplot2)
ggplot(simple_clusters, aes(x=X1, y=X2)) +
  geom_point() + theme(aspect.ratio=1)
}
\seealso{
clusters
}
\keyword{datasets}
