% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusters}
\alias{clusters}
\title{Three clusters in 5D}
\format{
A dataset with 300 rows and 6 columns
}
\description{
This data is simulated to use for testing.
It has three elliptical clusters in mostly
variables 2 and 4. They are not equidistant.
}
\details{
\describe{
\item{x1, x2, x3, x4, x5}{five numeric variables}
\item{cl}{class variable}
}
}
\examples{
clusters_pca <- prcomp(clusters[,1:5])
ggscree(clusters_pca)
}
\seealso{
simple_clusters
}
\keyword{datasets}
