% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mulgar-package.R
\docType{package}
\name{mulgar-package}
\alias{mulgar}
\alias{mulgar-package}
\title{mulgar: Functions for Pre-Processing Data for Multivariate Data Visualisation using Tours}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This is a companion to the book Cook, D. and Laa, U. (2023) \url{https://dicook.github.io/mulgar_book/} "Interactively exploring high-dimensional data and models in R". by Cook and Laa. It contains useful functions for processing data in preparation for visualising with a tour. There are also several sample data sets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dicook.github.io/mulgar/}
  \item \url{https://github.com/dicook/mulgar}
  \item Report bugs at \url{https://github.com/dicook/mulgar/issues}
}

}
\author{
\strong{Maintainer}: Dianne Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})

Authors:
\itemize{
  \item Ursula Laa \email{ursula.laa@boku.ac.at} (\href{https://orcid.org/0000-0002-0249-6439}{ORCID})
}

}
\keyword{internal}
