% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plane}
\alias{plane}
\title{2D plane in 5D}
\format{
A data set with 100 rows and 5 columns
}
\description{
This data is simulated to use for testing.
It has two dimensions of variability and
three of noise. It is created from a 2 factor
model, where all variables are related.
}
\details{
\describe{
\item{x1, x2, x3, x4, x5}{five numeric variables}
}
}
\examples{

plane_pca <- prcomp(plane)
ggscree(plane_pca)
}
\seealso{
box
}
\keyword{datasets}
